requirements.txt
python-telegram-bot==21.6
python-dotenv
loguru
pandas
matplotlib
numpy
TA-Lib
binance-asyncio
